third_party/WebKit/LayoutTests/editing/selection/caret-in-empty-inline-2-expected.txt

Bug 85793: Caret is not rendered in empty inline contenteditable elements

This test verifies that an empty inline contenteditable element, placed after another inline element, gets a valid caret rect.

Previous span

PASS caretRect.left is 268
PASS caretRect.top is 180
PASS caretRect.width is 1
PASS caretRect.height is 20
PASS successfullyParsed is true

TEST COMPLETE
